skoro to funguje kappa

+++ src/lexer.c
@@ -572,7 +572,7 @@ int getNextToken(token_t *token)
     if (token->type == T_String)
     {
         char *string = malloc(sizeof(char) * (strlen(token->data) - 1));
-        for (int i = 0; i < strlen(token->data) - 2; i++)
+        for (size_t i = 0; i < strlen(token->data) - 2; i++)
         {
             string[i] = token->data[i + 1];
         }
